Refactored CLI module.
We now split the input into smaller parts so that each core can work on
a separate part.  This speeds up splitting significantly on multi-core
machines.

Integrated plumbum for command-line work.

We no longer work with streams, because working with files is easier: we
know their size in advance, so it's easier to know if splitting is
necessary.  All our input is files anyway.